Finished the grades page
Finished the grades page so it would display a table to see and search grades, buttons to edit and delete grades, and button and page to create grade, only to Admins and Teachers. Updated grades page so Parents would see their children and then the grades of each children.

Ran build to deploy to render